% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df22-data.R
\docType{data}
\name{df22}
\alias{df22}
\title{A dataset of NEA Nuclear Energy Data 2018}
\format{A dataset of submitted results with 60 observations of 3 variables

\describe{
  \item{COUNTRY}{Country}
  \item{YEAR}{Year. If Year > 2018, values are 'expected'}
  \item{TU_PROD_PERYEAR}{rta}
}}
\usage{
data(df22)
}
\description{
The structure of this dataset (7 features per observation)
is the minimal template for calculated results
to be imported into the sendis package.
}
\keyword{datasets}
